Improve verify token

+++ Backend/controllers/message.js
@@ -1,11 +1,28 @@
 
 const db = require('../models');
+const jwt = require('jsonwebtoken');
 
 exports.createMessage = (req, res, next) => {
     const message = req.body;
-    db.Message.create(message)
-        .then(() => { res.status(201).json({ message: 'Message crée !' }) })
-        .catch(error => { res.status(500).json({ error }) })
+    const token = req.headers.authorization.split(' ')[1];
+    jwt.verify(token, process.env.TOKEN, function (err, decoded) {
+        if (err) {
+            err = {
+                name: 'TokenExpiredError',
+                message: 'jwt expired',
+            }
+        }
+        if (!decoded) {
+            console.log('invalid token')
+            res.status(500).json({ err });
+        } else {
+            db.Message.create(message)
+                .then(() => { res.status(201).json({ message: 'Message crée !' }) })
+                .catch(error => { res.status(500).json({ error }) })
+        }
+    })
+
+
 }
 
 exports.getAllMessages = (req, res, next) => {
@@ -14,6 +31,11 @@ exports.getAllMessages = (req, res, next) => {
             include: [{
                 model: db.User,
                 attributes: ['id', 'name'],
+
+            }, {
+                model: db.Comment,
+                as: 'Comments',
+                attributes: ['id']
 
             }],
             order: [['id', 'DESC']],
@@ -37,18 +59,40 @@ exports.getOneMessage = (req, res, next) => {
 }
 exports.modifyMessage = (req, res, next) => {
     const newMessage = req.body
-    console.log('newmessage', req.body)
-    db.Message.update(newMessage, { where: { id: req.query.id } })
-        .then(() => {
-            res.status(200).json({ message: 'Message modifié' })
+    const token = req.headers.authorization.split(' ')[1];
+    const decodedToken = jwt.verify(token, process.env.TOKEN);
+    const userId = decodedToken.userId;
+    db.Message.findOne({ where: { id: req.query.id } })
+        .then(message => {
+            if (message.user_id != userId) {
+                console.log('Token invalid')
+            } else {
+                db.Message.update(newMessage, { where: { id: req.query.id } })
+                    .then(() => {
+                        res.status(200).json({ message: 'Message modifié' })
+                    })
+                    .catch(error => res.status(500).json({ error }))
+            }
         })
-        .catch(error => console.log(error))
+        .catch(error => res.status(404).json({ error }))
 }
 
 exports.deleteMessage = (req, res, next) => {
-    db.Message.destroy({ where: { id: req.query.id } })
-        .then(() => {
-            res.status(204).json({ message: 'Message supprimé ' })
+    const token = req.headers.authorization.split(' ')[1];
+    const decodedToken = jwt.verify(token, process.env.TOKEN);
+    const userId = decodedToken.userId;
+    db.Message.findOne({ where: { id: req.query.id } })
+        .then(message => {
+            if (message.user_id != userId) {
+                console.log('Token invalid')
+            } else {
+                db.Message.destroy({ where: { id: req.query.id } })
+                    .then(() => {
+                        res.status(204).json({ message: 'Message supprimé ' })
+                    })
+                    .catch(error => res.status(500).json({ error }))
+            }
         })
         .catch(error => res.status(404).json({ error }))
+
 }
